Apply glyph and mark position adjustments
